Follow the Microsoft DNS client behaviour much more closely.
I have completely redone the update logic to match that documented in
the MMC help (Dns/Concepts/Understanding DNS/Dynamic update).

The client also reads configuration variables from the file
$(sysconfdir)/dnsupdate.conf if it exists. This is intended for
use with group policy.

+++ dnstcp.c
@@ -155,7 +155,8 @@ dnstcp_send(int s, const void *buf, size_t len)
 	return -1;
     }
     if (verbose > 3) 
-	fprintf(stderr, "dnstcp_send: writing %d bytes to fd %d\n", len, s);
+	fprintf(stderr, "dnstcp_send: writing %d bytes to fd %d\n", 
+		(int)len, s);
     if (len > 0 && write(s, buf, len) != len) {
 	warn("write");
 	return -1;
@@ -198,7 +199,8 @@ dnstcp_recv(int s, void *buf, size_t bufsz)
 	    return 0;
 	}
         if (verbose > 3)
-	   fprintf(stderr, "[read %d of %d header]\n", pos+len, sizeof b);
+	   fprintf(stderr, "[read %d of %d header]\n", pos+len, 
+		   (int)sizeof b);
     }
